+++ bitsv/network/services/bitindex.py
@@ -1,7 +1,6 @@
 import json
 
 import requests
-from cashaddress import convert as cashaddress
 
 from bitsv.network import currency_to_satoshi
 from bitsv.network.meta import Unspent
@@ -12,7 +11,6 @@ class BitIndex:
     @staticmethod
     def get_balance(address):
         """Gets utxos for given legacy address"""
-        address = cashaddress.to_legacy_address(address)
         headers = {
             'Content-Type': 'application/json',
             'Accept': 'application/json'
@@ -20,18 +18,19 @@ class BitIndex:
 
         r = requests.get('https://api.bitindex.network/api/v2/addrs/balance?address={}'.format(address),
                          headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
     def get_txs(address):
         """Gets utxos for given legacy address"""
-        address = cashaddress.to_legacy_address(address)
         headers = {
             'Content-Type': 'application/json',
             'Accept': 'application/json'
         }
 
         r = requests.get('https://api.bitindex.network/api/v2/addrs/txs?address={}'.format(address), headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
@@ -43,12 +42,12 @@ class BitIndex:
         }
 
         r = requests.get('https://api.bitindex.network/api/v2/tx/{}'.format(tx), headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
     def get_utxo(address):
         """gets utxos for given address BitIndex api"""
-        address = cashaddress.to_legacy_address(address)
         json_payload = json.dumps({"addrs": address})
 
         headers = {
@@ -56,6 +55,7 @@ class BitIndex:
             'Accept': 'application/json'
         }
         r = requests.post('https://api.bitindex.network/api/addrs/utxo', data=json_payload, headers=headers)
+        r.raise_for_status()
         return [Unspent(amount=currency_to_satoshi(tx['amount'], 'bsv'),
                         script=tx['scriptPubKey'],
                         txid=tx['txid'],
@@ -72,6 +72,7 @@ class BitIndex:
             'Accept': 'application/json'
         }
         r = requests.post('https://api.bitindex.network/api/v2/tx/send', data=json_payload, headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
@@ -94,6 +95,7 @@ class BitIndex:
         }
         r = requests.get('https://api.bitindex.network/api/v2/xpub/status?xpub={}'.format(xpub),
                          headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
@@ -105,6 +107,7 @@ class BitIndex:
         }
         r = requests.get('https://api.bitindex.network/api/v2/xpub/balance?xpub={}'.format(xpub),
                          headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
@@ -116,6 +119,7 @@ class BitIndex:
         }
         r = requests.get('https://api.bitindex.network/api/v2/xpub/utxos?xpub={}'.format(xpub),
                          headers=headers)
+        r.raise_for_status()
         return r.json()
 
     @staticmethod
@@ -127,4 +131,5 @@ class BitIndex:
         }
         r = requests.get('https://api.bitindex.network/api/v2/xpub/txs?xpub={}'.format(xpub),
                          headers=headers)
+        r.raise_for_status()
         return r.json()